[ios] prep, break out pool handling and add path/framework path handing
