+++ xbmc/Util.cpp
@@ -70,7 +70,11 @@
 #include "WIN32Util.h"
 #endif
 #if defined(__APPLE__)
-#include "CocoaInterface.h"
+#if defined(__arm__)
+#include "osx/iOSUtils.h"
+#else
+#include "osx/CocoaInterface.h"
+#endif
 #endif
 #include "GUIUserMessages.h"
 #include "filesystem/File.h"
@@ -494,15 +498,23 @@ void CUtil::GetHomePath(CStdString& strPath, const CStdString& strTarget)
     char     given_path[2*MAXPATHLEN];
     uint32_t path_size = 2*MAXPATHLEN;
 
-    result = _NSGetExecutablePath(given_path, &path_size);
+    #if defined(__arm__)
+      result = GetIOSExecutablePath(given_path, &path_size);
+    #else
+      result = _NSGetExecutablePath(given_path, &path_size);
+    #endif
     if (result == 0)
     {
       // Move backwards to last /.
       for (int n=strlen(given_path)-1; given_path[n] != '/'; n--)
         given_path[n] = '\0';
 
-      // Assume local path inside application bundle.
-      strcat(given_path, "../Resources/XBMC/");
+      #if defined(__arm__)
+        strcat(given_path, "/XBMCData/XBMCHome/");
+      #else
+        // Assume local path inside application bundle.
+        strcat(given_path, "../Resources/XBMC/");
+      #endif
 
       // Convert to real path.
       char real_path[2*MAXPATHLEN];
@@ -2165,7 +2177,11 @@ CStdString CUtil::ResolveExecutablePath()
   char     real_given_path[2*MAXPATHLEN];
   uint32_t path_size = 2*MAXPATHLEN;
 
-  result = _NSGetExecutablePath(given_path, &path_size);
+  #if defined(__arm__)
+    result = GetIOSExecutablePath(given_path, &path_size);
+  #else
+    result = _NSGetExecutablePath(given_path, &path_size);
+  #endif
   if (result == 0)
     realpath(given_path, real_given_path);
   strExecutablePath = real_given_path;
@@ -2189,30 +2205,29 @@ CStdString CUtil::GetFrameworksPath(void)
 {
   CStdString strFrameworksPath;
 #if defined(__APPLE__)
-  int      result = -1;
   char     given_path[2*MAXPATHLEN];
   char     real_given_path[2*MAXPATHLEN];
   uint32_t path_size = 2*MAXPATHLEN;
 
-  result = _NSGetExecutablePath(given_path, &path_size);
-  if (result == 0)
-  {
+  #if defined(__arm__)
+    GetIOSFrameworkPath(given_path, &path_size);
+  #else
+    _NSGetExecutablePath(given_path, &path_size);
     // Move backwards to last /.
     for (int n=strlen(given_path)-1; given_path[n] != '/'; n--)
       given_path[n] = '\0';
 
     // Assume local path inside application bundle.
     strcat(given_path, "../Frameworks");
-
-    // Convert to real path.
-    if (realpath(given_path, real_given_path) != NULL)
+  #endif
+  // Convert to real path.
+  if (realpath(given_path, real_given_path) != NULL)
+  {
+    // check if we are running as real xbmc.app
+    if (strstr(real_given_path, "Contents"))
     {
-      // check if we are running as real xbmc.app
-      if (strstr(real_given_path, "Contents"))
-      {
-        strFrameworksPath = real_given_path;
-        return strFrameworksPath;
-      }
+      strFrameworksPath = real_given_path;
+      return strFrameworksPath;
     }
   }
 #endif
