+++ xbmc/osx/iOSUtils.h
@@ -1,0 +1,41 @@
+/*
+ *      Copyright (C) 2010 Team XBMC
+ *      http://www.xbmc.org
+ *
+ *  This Program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2, or (at your option)
+ *  any later version.
+ *
+ *  This Program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with XBMC; see the file COPYING.  If not, write to
+ *  the Free Software Foundation, 675 Mass Ave, Cambridge, MA 02139, USA.
+ *  http://www.gnu.org/copyleft/gpl.html
+ *
+ */
+#ifndef _IOS_UTILS_H_
+#define _IOS_UTILS_H_
+
+#include <string>
+#include <CoreFoundation/CFURL.h>
+#import "osx/AutoPool.h"
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+  CFURLRef CreateCFURLRefFromFilePath(const char *filepath);
+
+  int     GetIOSFrameworkPath(char* path, uint32_t *pathsize);
+  int     GetIOSExecutablePath(char* path, uint32_t *pathsize);
+  bool    iOS_HasVideoToolboxDecoder(void);
+#ifdef __cplusplus
+}
+#endif
+
+#endif
